+++ src/resourceManager.ts
@@ -408,6 +408,24 @@ export class KicCell extends EventEmitter {
             })
         }
 
+        vscode.window.onDidCloseTerminal((t) => {
+            if (
+                t.creationOptions.iconPath !== undefined &&
+                t.creationOptions.iconPath.toString().search("keithley-logo") &&
+                t.exitStatus !== undefined &&
+                t.exitStatus.reason !== vscode.TerminalExitReason.Process
+            ) {
+                setTimeout(() => {
+                    child.spawnSync(EXECUTABLE, [
+                        "-v",
+                        "reset",
+                        connType,
+                        unique_id,
+                    ])
+                }, 500)
+            }
+        })
+
         this.terminalPid = this._term.processId
 
         if (this._term != undefined) {