+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 - Reset instrument if closed unexpectedly using the trashcan on the terminal (TSP-730)
 - Add logging for terminal and discover
+- Default friendly name if user doesn't provide one (TSP-757)
 
 ## [0.16.4]