+++ CHANGELOG.md
@@ -16,6 +16,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     Security -- in case of vulnerabilities.
 -->
 
+## [0.17.1]
+
+### Fixed
+- Succesful connection to tspop adds the instrument details to Instruments pane (TSP-773)
+
 ## [0.17.0]
 
 ### Added